input/dialog split
