+++ selfdrive/ui/qt/widgets/input.h
@@ -1,25 +1,10 @@
 #pragma once
 
-#include <QDialog>
-#include <QLabel>
 #include <QLineEdit>
 #include <QString>
-#include <QVBoxLayout>
-#include <QWidget>
 
+#include "selfdrive/ui/qt/widgets/dialog.h"
 #include "selfdrive/ui/qt/widgets/keyboard.h"
-
-
-class DialogBase : public QDialog {
-  Q_OBJECT
-
-protected:
-  DialogBase(QWidget *parent);
-  bool eventFilter(QObject *o, QEvent *e) override;
-
-public slots:
-  int exec() override;
-};
 
 class InputDialog : public DialogBase {
   Q_OBJECT
@@ -48,24 +33,4 @@ private slots:
 signals:
   void cancel();
   void emitText(const QString &text);
-};
-
-class ConfirmationDialog : public DialogBase {
-  Q_OBJECT
-
-public:
-  explicit ConfirmationDialog(const QString &prompt_text, const QString &confirm_text,
-                              const QString &cancel_text, const bool rich, QWidget* parent);
-  static bool alert(const QString &prompt_text, QWidget *parent);
-  static bool confirm(const QString &prompt_text, const QString &confirm_text, QWidget *parent);
-  static bool rich(const QString &prompt_text, QWidget *parent);
-};
-
-class MultiOptionDialog : public DialogBase {
-  Q_OBJECT
-
-public:
-  explicit MultiOptionDialog(const QString &prompt_text, const QStringList &l, const QString &current, QWidget *parent);
-  static QString getSelection(const QString &prompt_text, const QStringList &l, const QString &current, QWidget *parent);
-  QString selection;
 };
